+++ scripts/bin/_env.sh
@@ -33,7 +33,7 @@ export OS=$(getOS;);
 export WEB_SERVER_PORT=1841;
 export AUTHOR="bradyhouse@gmail.com";
 export AUTHOR_NAME="brady house";
-export NVM_VERSION="v10.9.0";
+export NVM_VERSION="v15.6.0";
 export CHANGE_LOG="CHANGELOG.md";
 
 # =================================================